2015-07-18,Add the function of counting.

+++ project/usr/inc/delay.h
@@ -3,7 +3,7 @@
 
 #include "stm8l15x.h"
 
-void delay1ms(uint8_t ms);
-void delay1s(uint8_t seconds);
+void delay1ms(uint16_t ms);
+void delay1s(uint16_t seconds);
 
 #endif
